test: run the shared scaffolding's self-tests once instead of 43 times
`tests/common/` is a module, not a crate: every integration test that
needs a WAV loader, a channel model or the golden helpers writes
`mod common;`, and 43 of them do. A `#[cfg(test)] mod tests` inside
`common/channel.rs` is therefore compiled and executed once per
including binary — 23 tests × 43 binaries ≈ 990 redundant executions
per `cargo test`.

Wall-clock cost was negligible, which is why this sat unfixed. The
real cost is that it made the pass count meaningless: ~750 of the
suite's reported 1374 "passed" were the same 23 tests over and over,
implying far more coverage than existed. A four-digit number that
mostly counts itself is worse than a small honest one, particularly
in a suite whose problem this whole restructure was that it looked
better covered than it was.

Collected into `tests/common_selftest.rs`, which owns `mod common;`
and runs those 23 once. The helpers stay where they are; only their
tests moved. Three private helpers they exercise
(`corpus::require_enabled`, `air_channel::{analytic_signal, bpf}`)
plus `air_channel::SAMPLE_RATE_HZ` were widened to `pub` — `common` is
test-only scaffolding, so that costs nothing.

Six `uvpacket_*` tests turned out to declare `mod common;` **without**
`#[allow(dead_code)]`, unlike the other 37. They compiled only because
`common`'s own `#[cfg(test)]` blocks were keeping several helpers
"used"; removing those exposed it. Fixed by adding the attribute those
six were missing, rather than by putting a blanket
`#![allow(dead_code)]` inside `common/mod.rs` — the first attempt did
exactly that and clippy's `duplicated_attributes` correctly rejected
it, which is also what proved the outer attribute already reaches the
submodules.

Result:

    total passed   1374  ->  635      (same coverage, honestly counted)
    binaries         88  ->   79

Full suite under `MFSK_REQUIRE_CORPUS=1` green;
`scripts/pre-push-check.sh` clean.
