+++ mfsk-core/tests/common/channel.rs
@@ -205,122 +205,3 @@ impl RayleighFlatChannel {
         ((self.state >> 32) as f32 + 1.0) / 4_294_967_297.0
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    /// AWGN channel with σ = 0 must be a no-op (modulo the f32
-    /// representation of `0.0 × x = 0.0`).
-    #[test]
-    fn awgn_with_zero_sigma_is_identity() {
-        let mut c = AwgnChannel::new(0.0, 42);
-        let mut buf = vec![0.5_f32; 100];
-        let original = buf.clone();
-        c.apply(&mut buf);
-        assert_eq!(buf, original);
-    }
-
-    /// AWGN with a non-zero σ produces non-zero noise — and seeded
-    /// runs are reproducible.
-    #[test]
-    fn awgn_seeded_is_reproducible() {
-        let mut a = AwgnChannel::new(0.5, 123);
-        let mut b = AwgnChannel::new(0.5, 123);
-        let mut buf_a = vec![0.0_f32; 200];
-        let mut buf_b = vec![0.0_f32; 200];
-        a.apply(&mut buf_a);
-        b.apply(&mut buf_b);
-        assert_eq!(buf_a, buf_b);
-        // …and at least 90 % of samples are non-zero.
-        let nonzero = buf_a.iter().filter(|&&s| s != 0.0).count();
-        assert!(nonzero > 180);
-    }
-
-    /// At a high Eb/N0 (= 100 dB) the σ for every mode should be
-    /// vanishingly small; at 0 dB it should be a meaningful fraction
-    /// of the signal envelope.
-    #[test]
-    fn sigma_decreases_with_eb_n0() {
-        // Use a representative 4-FSK-era signal power (P = 0.5) so
-        // the test bounds are absolute rather than measurement-tied.
-        let p = 0.5;
-        for mode in [
-            Mode::Robust,
-            Mode::Standard,
-            Mode::UltraRobust,
-            Mode::Express,
-        ] {
-            let sigma_clean = awgn_sigma_for_eb_n0_info(mode, 100.0, p);
-            let sigma_zero = awgn_sigma_for_eb_n0_info(mode, 0.0, p);
-            assert!(sigma_clean < 1e-3, "{mode:?}: clean σ {sigma_clean}");
-            assert!(sigma_zero > 0.5, "{mode:?}: 0-dB σ {sigma_zero}");
-            assert!(
-                sigma_clean < sigma_zero,
-                "{mode:?}: σ should decrease as Eb/N0 grows",
-            );
-        }
-    }
-
-    /// Higher-rate modes spend less channel energy per info bit, so
-    /// at a fixed Eb/N0_info the channel-domain noise must be lower
-    /// for them — i.e. σ decreases with rate.
-    #[test]
-    fn sigma_decreases_with_rate() {
-        // UltraRobust shares Robust's FEC rate (only the symbol
-        // rate differs), so σ at fixed Eb/N0 is identical between
-        // them — exclude UltraRobust from this strict-monotonic
-        // test and check only the FEC-rate-distinct modes.
-        let eb_n0 = 0.0;
-        let p = 0.5;
-        let sigmas: Vec<f32> = [Mode::Robust, Mode::Standard, Mode::Express]
-            .iter()
-            .map(|&m| awgn_sigma_for_eb_n0_info(m, eb_n0, p))
-            .collect();
-        for w in sigmas.windows(2) {
-            assert!(
-                w[0] > w[1],
-                "expected σ to decrease across rates: {sigmas:?}",
-            );
-        }
-    }
-
-    /// σ scales as `√P` (since variance scales as P at fixed Eb/N0).
-    #[test]
-    fn sigma_scales_with_signal_power() {
-        let s1 = awgn_sigma_for_eb_n0_info(Mode::Robust, 0.0, 1.0);
-        let s4 = awgn_sigma_for_eb_n0_info(Mode::Robust, 0.0, 4.0);
-        // 4× power → 2× σ (within float epsilon).
-        assert!((s4 / s1 - 2.0).abs() < 1e-3, "s1={s1}, s4={s4}");
-    }
-
-    /// Rayleigh envelope statistics: over a long buffer, the
-    /// magnitude has E[|h|²] ≈ 1 by construction.
-    #[test]
-    fn rayleigh_envelope_has_unit_mean_square() {
-        let mut chan = RayleighFlatChannel::new(5.0, 0.0, 0xABC1_2345);
-        let mut audio = vec![1.0_f32; 12_000]; // 1 sec at fs=12kHz
-        chan.apply(&mut audio);
-        let mean_sq: f32 = audio.iter().map(|s| s * s).sum::<f32>() / audio.len() as f32;
-        // Allow generous tolerance because 1 sec at 5 Hz Doppler
-        // ≈ 5 fading cycles → high finite-sample variance.
-        assert!(
-            (0.5..2.0).contains(&mean_sq),
-            "Rayleigh mean-square {mean_sq} far off the expected 1.0",
-        );
-    }
-
-    /// Rayleigh + AWGN at huge Doppler approaches AWGN-only stats
-    /// in the limit (envelope decorrelates fast, magnitude
-    /// distribution converges to Rayleigh independent of past).
-    /// Sanity smoke test that nothing panics or NaNs.
-    #[test]
-    fn rayleigh_apply_does_not_nan() {
-        for &fd in &[0.5_f32, 1.0, 5.0, 20.0] {
-            let mut chan = RayleighFlatChannel::new(fd, 0.1, 0xFEED + fd as u64);
-            let mut audio = vec![0.5_f32; 4096];
-            chan.apply(&mut audio);
-            assert!(audio.iter().all(|s| s.is_finite()));
-        }
-    }
-}
