+++ mfsk-core/tests/common/golden.rs
@@ -213,103 +213,3 @@ pub fn assert_golden<T>(
         snr_errors.join("\n  ")
     );
 }
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    static EXPECTED: &[GoldenEntry] = &[
-        GoldenEntry {
-            msg: "CQ K1ABC FN42",
-            freq_hz: Some(1000.0),
-            dt_sec: Some(0.0),
-            snr_db: Some(-10.0),
-        },
-        GoldenEntry {
-            msg: "K1ABC W9XYZ EN37",
-            freq_hz: None,
-            dt_sec: None,
-            snr_db: None,
-        },
-    ];
-
-    fn set(min_hits: usize, max_extra: usize) -> GoldenSet {
-        GoldenSet {
-            name: "test",
-            expected: EXPECTED,
-            min_hits,
-            max_extra,
-        }
-    }
-
-    fn v(msg: &str, freq_hz: f32, snr_db: Option<f32>) -> DecodeView {
-        DecodeView {
-            msg: msg.to_string(),
-            freq_hz,
-            dt_sec: 0.0,
-            snr_db,
-        }
-    }
-
-    #[test]
-    fn passes_when_recall_and_precision_both_hold() {
-        let d = [
-            v("CQ K1ABC FN42", 1000.0, Some(-11.0)),
-            v("K1ABC W9XYZ EN37", 1500.0, None),
-        ];
-        assert_golden(&d, &set(2, 0), Tolerances::default(), |x| x.clone());
-    }
-
-    /// The regression that motivated this module: full recall, but the
-    /// decoder also invented something.
-    #[test]
-    #[should_panic(expected = "outside the golden set")]
-    fn full_recall_does_not_excuse_a_phantom() {
-        let d = [
-            v("CQ K1ABC FN42", 1000.0, Some(-10.0)),
-            v("K1ABC W9XYZ EN37", 1500.0, None),
-            v("UZC/7D0DKY 17", 1501.0, None),
-        ];
-        assert_golden(&d, &set(2, 0), Tolerances::default(), |x| x.clone());
-    }
-
-    #[test]
-    #[should_panic(expected = "recall regressed")]
-    fn missing_entry_fails_recall() {
-        let d = [v("CQ K1ABC FN42", 1000.0, Some(-10.0))];
-        assert_golden(&d, &set(2, 0), Tolerances::default(), |x| x.clone());
-    }
-
-    #[test]
-    #[should_panic(expected = "reported SNR is out of tolerance")]
-    fn snr_outside_tolerance_fails() {
-        let d = [
-            v("CQ K1ABC FN42", 1000.0, Some(22.0)),
-            v("K1ABC W9XYZ EN37", 1500.0, None),
-        ];
-        assert_golden(&d, &set(2, 0), Tolerances::default(), |x| x.clone());
-    }
-
-    /// A frequency far from the golden entry must not count as a hit,
-    /// or the freq column is decorative.
-    #[test]
-    #[should_panic(expected = "recall regressed")]
-    fn wrong_frequency_is_not_a_hit() {
-        let d = [
-            v("CQ K1ABC FN42", 1400.0, Some(-10.0)),
-            v("K1ABC W9XYZ EN37", 1500.0, None),
-        ];
-        assert_golden(&d, &set(2, 9), Tolerances::default(), |x| x.clone());
-    }
-
-    #[test]
-    fn documented_budgets_are_honoured() {
-        // A tracked gap (min_hits below len) and a tracked phantom
-        // budget both pass without weakening the other assertion.
-        let d = [
-            v("CQ K1ABC FN42", 1000.0, Some(-10.0)),
-            v("SOMETHING ELSE", 1600.0, None),
-        ];
-        assert_golden(&d, &set(1, 1), Tolerances::default(), |x| x.clone());
-    }
-}
