+++ mfsk-core/tests/uvpacket_multichannel.rs
@@ -9,6 +9,13 @@ use mfsk_core::uvpacket::framing::FrameHeader;
 use mfsk_core::uvpacket::rx::{MultiChannelOpts, decode_multichannel, measure_slot_energies};
 use mfsk_core::uvpacket::{Mode, tx};
 
+// `common` is shared scaffolding included by 43 test binaries, each
+// using a different subset — dead-code analysis of it is per-binary
+// noise. Every other consumer already carries this attribute; these
+// six only compiled without it because `common`'s own `#[cfg(test)]`
+// blocks kept the helpers "used" until they moved to
+// `tests/common_selftest.rs`.
+#[allow(dead_code)]
 mod common;
 use common::channel::{AwgnChannel, awgn_sigma_for_eb_n0_info, signal_power};
